+++ src/main/java/br/pucrio/les/esg_token_backend/models/Value.java
@@ -1,7 +1,5 @@
 package br.pucrio.les.esg_token_backend.models;
 
-import java.util.Date;
-
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
@@ -9,80 +7,22 @@ import javax.validation.constraints.Size;
 
 @Entity
 @Table(name = "value")
-public class Value {
-    @Id
-    @Column
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+public class Value extends BaseModel {
     @Column
     @NotEmpty(message = "Can not be empty")
     @NotBlank(message = "Can not be blank")
     @Size(min = 3, max = 255)
     private String value;
 
-    @Column
-    private Date createdAt;
-
-    @Column
-    private Date updatedAt;
-
     public Value() {
-    }
-
-    public Value(String value) {
-        this.value = value;
-    }
-
-    public Value(String value, Date createdAt, Date updatedAt) {
-        this.value = value;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
     }
 
     public void setValue(String value) {
         this.value = value;
     }
 
-    public void setId(Long id) {
-        this.id = id;
-    }
-
     public String getValue() {
         return value;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public Date getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(Date updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-
-    @PrePersist
-    public void onPrePersist() {
-        if (this.createdAt == null) {
-            this.createdAt = new Date();
-            this.updatedAt = this.createdAt;
-        }
-    }
-
-    @PreUpdate
-    public void onPreUpdate() {
-        this.updatedAt = new Date();
     }
 
 }